fix: deleting clothing size

+++ src/main/java/app/service/controllers/admin/shop/clothing/impl/ClothingSizeServiceImpl.java
@@ -37,7 +37,9 @@ public class ClothingSizeServiceImpl implements ClothingSizeService {
 
     @Override
     public void deleteById(Integer id) {
-        clothingSizeRepository.deleteById(id);
+        if (clothingSizeRepository.existsById(id)) {
+            clothingSizeRepository.deleteById(id);
+        }
     }
 
     @Override
